+++ SimpleClicker/changelog.md
@@ -1,4 +1,11 @@
-﻿## Version 0.9.1.0:
+﻿## Version 0.9.1.1 (QOL):
+	- Changes tab replacement (Interfaces <-> Extras)
+	- Changes a few UI elements.
+	- Changes some default options.
+	- Added more tooltip descriptions for question boxes.
+	- Color options will now show its value in its ribbon.
+
+## Version 0.9.1.0:
 	- Replaces some default options' placements.
 	- Added new menu and more options.
 	- All settings will now be saved immediately.